distribution_service/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.2
django-allauth==0.61.1
django-filter==23.5
django-redis==5.4.0
django-rq==2.10.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
oauthlib==3.2.2
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
rq==1.15.1
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.0
whitenoise==6.6.0
